Исправлена ошибка с подключением к БД,

+++ SuterShop/CentralPanel/ViewModel/CenterPanelViewModel.cs
@@ -37,8 +37,6 @@ namespace SuterShop.CentralPanel.View
         internal void SetData()
         {
             _db = (Application.Current as IApp).Db;
-            _db.Database.CloseConnection();
-            _db.Database.OpenConnection();
             var goods = _db.GoodsForSaleList.Include("User").ToList();
             // Получаем папку куда пользователь установил нашу программу.
             var dir = $"{Directory.GetCurrentDirectory()}{System.IO.Path.DirectorySeparatorChar}TempImages{System.IO.Path.DirectorySeparatorChar}";
@@ -51,7 +49,10 @@ namespace SuterShop.CentralPanel.View
             foreach (var good in goods)
             {
                 var fileName = $"{good.Id}_{good.Name}.png";
-                File.WriteAllBytes($"{dir}{fileName}", good.Image);
+                if (!File.Exists($"{dir}{fileName}"))
+                {
+                    File.WriteAllBytes($"{dir}{fileName}", good.Image);
+                }
                 var card = new cardView();
                 card.Margin = new Thickness(5);
                 (card.DataContext as cardViewModel).SetData(goods, good, $"{dir}{fileName}");
